+++ PermissionServer/PermissionServer.cs
@@ -25,9 +25,22 @@ namespace PermissionServer
         /// store a token OR the failure to send it. For security reasons the
         /// specifics are NOT returned as the end user should never be told.
         /// </summary>
-        public async Task<bool> StartConfirmation(string emailAddress, string confirmationUrl)
+        /// <param name="context">
+        /// Passing in a context will tie the generated token to the value
+        /// provided, effectively meaning it will only be matched if confirmed
+        /// from the same context.  Leave it off to allow tokens to be confirmed
+        /// in any context.
+        /// 
+        /// You could use an IP address for a web site. Another option is a
+        /// browser 'fingerprint', or a GUID from a secure cookie works well and
+        /// restricts to the same session not just context.
+        /// </param>
+        public async Task<bool> StartConfirmation(
+            string emailAddress,
+            string confirmationUrl,
+            string context = "")
         {
-            var token = tokenStore.Add(emailAddress);
+            var token = tokenStore.Add(emailAddress, context);
             if (token == null) return false;
             (var ok, var error) = await mailer.SendConfirmation(emailAddress, token, confirmationUrl);
             if (!ok)
@@ -43,9 +56,20 @@ namespace PermissionServer
         /// address (and has not expired). If it doesn't then the
         /// return value states why.
         /// </summary>
-        public bool CompleteConfirmation(string emailAddress, string confirmationCode)
+        /// <param name="context">
+        /// If you generated the token including a context then
+        /// it will only match if the same context is also passed
+        /// in here, effectively restricting token confirmations to
+        /// the same context that issued them (for extra security).
+        /// If no context was used during token generation, leave
+        /// it off here.
+        /// </param>
+        public bool CompleteConfirmation(
+            string emailAddress,
+            string confirmationCode,
+            string context = "")
         {
-            var token = tokenStore.Get(emailAddress, confirmationCode);
+            var token = tokenStore.Get(emailAddress, confirmationCode, context);
             if (token == null) return false;
             if (options.Tokens.SingleUse) tokenStore.Remove(emailAddress, confirmationCode);
             return true;