+++ PermissionServer/Token.cs
@@ -9,14 +9,23 @@ namespace PermissionServer
         public DateTime CreatedAt { get; }
         public DateTime ValidUntil { get; }
         public int LifetimeMinutes { get; }
+        public string Context { get; set; }
 
         public string ConfirmationCode { get; }
 
-        /// <summary>Returns true if the token has expired.</summary>
+        /// <summary>True if the token has expired.</summary>
         public bool IsExpired => ValidUntil < DateTime.UtcNow;
 
-        /// <summary>Returns true if the token hasn't expired.</summary>
+        /// <summary>True if the token hasn't expired.</summary>
         public bool IsActive => IsExpired == false;
+
+        /// <summary>True if the token has a populated Context.</summary>
+        /// <remarks>
+        /// A context-specific token should only match if the same context
+        /// is passed to comparisons (a null/empty Context should match
+        /// regardless).
+        /// </remarks>
+        public bool HasContext => Context.HasValue();
 
         /// <summary>
         /// Creates a token with the specified length and lifetime.
@@ -24,12 +33,13 @@ namespace PermissionServer
         /// Regardless of the tokenLength, a minimum of 5 digits is
         /// enforced.
         /// </summary>
-        public Token(int tokenLength, int lifetimeMinutes)
+        public Token(int tokenLength, int lifetimeMinutes, string context)
         {
             CreatedAt = DateTime.UtcNow;
             ValidUntil = CreatedAt.AddMinutes(lifetimeMinutes);
             ConfirmationCode = GetTokenString(tokenLength);
             LifetimeMinutes = lifetimeMinutes;
+            Context = context;
         }
 
         /// <summary>
@@ -41,14 +51,20 @@ namespace PermissionServer
 
         /// <summary>
         /// Returns true if the (case-insensitive) confirmation
-        /// code matches.
+        /// code matches. If the token is context-specific then
+        /// the passed-in context is also checked.
         /// </summary>
-        public bool Matches(string suggestedConfirmationCode)
+        public bool Matches(string suggestedConfirmationCode, string context = "")
         {
-            return string.Equals(
+            var contextOK = true;
+            if (HasContext) contextOK = (context == Context);
+
+            var codeOK = string.Equals(
                 ConfirmationCode,
                 suggestedConfirmationCode.Trim(),
                 StringComparison.OrdinalIgnoreCase);
+
+            return contextOK && codeOK;
         }
 
         private string GetTokenString(int tokenDigits)